HTTPConnection: detect closed sockets during read phase of request()
If the server closed the socket after we sent our request but
explicitly requested keepalive, our end of the socket may appear
readable in select(). If this happens, not only log that the socket
appears closed in read, but close the socket and open a new one. This
was observed in the wild after a large changegroupsubset request from
a Mercurial server behind nginx. Includes a test.