request: make sure response bytes are arriving before assuming request worked
There's a risk when reusing a socket that we'll find out the server
closed the socket after sending the most recent response, but that we
won't see that RST on the socket until we've already sent an entire
request. This change defends against that by waiting for *any* amount
of response traffic after sending a request, since at that point we
wouldn't be safe to transparently re-send the request anyway.

Introduces a test with an intentionally-malicious TCP server that
simulates this kind of latency. It was easier to write a test using a
real socket than to try and further complicate mocksocket for this
subtle case.