proxy support: add proxy header support for CONNECT-based proxied requests
This adds two features, which are intermingled. The first is the
ability to specify a Proxy-Authorization header in the regular
request() codepath, which is compatible with httplib.

In the interest of completeness, proxy headers may also be specified
at HTTPConnection construction time, in which case those headers will
be preserved for the life of the HTTPConnection and passed to any
CONNECT-based proxy requests.